requirements.txt
fastapi
uvicorn[standard]
httpx[http2]
structlog
python-dotenv
selenium>=4.15.0
webdriver-manager
boto3
